% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_zeros.R
\name{count_zeros}
\alias{count_zeros}
\title{count_zeros}
\usage{
count_zeros(x)
}
\arguments{
\item{x}{A vector}
}
\value{
numeric vector of length 1 giving number of zero values in vector \code{x}.
}
\description{
\code{count_zeros} Count number of zeros within a vector
}
\details{
This functions counts the number of zeros in a vector
}
\examples{
a <- cbind(1:10,0)
count_zeros(x)
}

