% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_plots.R
\name{make_plot}
\alias{make_plot}
\title{make_plot}
\usage{
make_plot(x)
}
\arguments{
\item{x}{A vector}
}
\value{
numeric vector of length 1 giving number of zero values in vector \code{x}.
}
\description{
\code{make_plot} Count number of zeros within a vector
}
\details{
This functions counts the number of zeros in a vector
}
\examples{
a <- cbind(1:10,0)
make_plot(x)
}

