% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/se_lines.R
\name{se_lines}
\alias{se_lines}
\title{se_lines}
\usage{
se_lines(x, SE, y, bar_width = 0.1)
}
\arguments{
\item{x}{A vector}
}
\value{
numeric vector of length 1 giving number of zero values in vector \code{x}.
}
\description{
\code{se_lines} Add s confidence intervals to scatterplot
}
\details{
This functions counts the number of zeros in a vector
}
\examples{
a <- cbind(1:10,0)
count_zeros(x)
}

