% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sent_case.R
\name{sent_case}
\alias{sent_case}
\title{sent_case}
\usage{
sent_case(x)
}
\arguments{
\item{x}{A vector}
}
\value{
numeric vector of length 1 giving number of zero values in vector \code{x}.
}
\description{
\code{sent_case} Count number of zeros within a vector
}
\details{
This functions counts the number of zeros in a vector
}
\examples{
a <- cbind(1:10,0)
make_plot(x)
}

